피드백 : 로또결과 관련 로직을 담당하는 LottoResults 클래스를 생성하여 로직 위임
