+++ src/main/java/lotto/util/LottoUtils.java
@@ -1,10 +1,10 @@
 package lotto.util;
 
 import java.util.Arrays;
-import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import lotto.domain.LottoResults;
 import lotto.domain.enums.LottoCorrectCount;
 import lotto.dto.LottoResult;
 
@@ -22,20 +22,10 @@ public class LottoUtils {
         return Arrays.stream(numbers).mapToObj(String::valueOf).toArray(String[]::new);
     }
 
-    public static Map<LottoCorrectCount, Integer> resultToMap(List<LottoResult> lottoResults) {
-        return lottoResults.stream()
+    public static Map<LottoCorrectCount, Integer> resultToMap(LottoResults lottoResults) {
+        return lottoResults.getLottoResults().stream()
                 .filter(lottoResult -> lottoResult.getCorrectCount() >= THREE.getValue())
                 .collect(groupingBy(LottoUtils::makeCorrectKey, Collectors.summingInt(LottoResult::sumCount)));
-    }
-
-    public static int getProfit(Map<LottoCorrectCount, Integer> resultMap) {
-        return resultMap.entrySet().stream()
-                .mapToInt(map -> map.getKey().getProfit() * map.getValue())
-                .sum();
-    }
-
-    public static float getProfitRate(int totalProfit, int money) {
-        return (float) totalProfit * 100 / (float) money;
     }
 
     private static LottoCorrectCount makeCorrectKey(LottoResult result) {
